+++ CHANGELOG.md
@@ -1,2 +1,3 @@
 - Fix demo projects + web frameworks with emulators (#6737)
+- Fix Next.js static routes with server actions (#6664)
 - Remove development-only files from Next.js deployments (#6731)